requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-common-compat
apache-airflow-providers-apache-spark
apache-airflow-providers-http
minio
requests
